make documentation comments behaviour part of cstyle; add option for auto-closing doc comments
