+++ src/mode/asl.js
@@ -4,12 +4,12 @@ var oop = require("../lib/oop");
 var TextMode = require("./text").Mode;
 var ASLHighlightRules = require("./asl_highlight_rules").ASLHighlightRules;
 var FoldMode = require("./folding/cstyle").FoldMode;
-var DocCommentBehaviour = require("../mode/behaviour/doc_comment").DocCommentBehaviour;
+var CstyleBehaviour = require("../mode/behaviour/cstyle").CstyleBehaviour;
 
 var Mode = function () {
     this.HighlightRules = ASLHighlightRules;
     this.foldingRules = new FoldMode();
-    this.$behaviour = new DocCommentBehaviour();
+    this.$behaviour = new CstyleBehaviour({closeDocComment: true});
 };
 oop.inherits(Mode, TextMode);
