+++ src/mode/edifact.js
@@ -3,12 +3,12 @@
 var oop = require("../lib/oop");
 var TextMode = require("./text").Mode;
 var EdifactHighlightRules = require("./edifact_highlight_rules").EdifactHighlightRules;
-var DocCommentBehaviour = require("../mode/behaviour/doc_comment").DocCommentBehaviour;
+var CstyleBehaviour = require("../mode/behaviour/cstyle").CstyleBehaviour;
 
 var Mode = function() {
    
     this.HighlightRules = EdifactHighlightRules;
-    this.$behaviour = new DocCommentBehaviour();
+    this.$behaviour = new CstyleBehaviour({closeDocComment: true});
 };
 oop.inherits(Mode, TextMode);
