+++ src/mode/jsx.js
@@ -4,13 +4,13 @@ var oop = require("../lib/oop");
 var TextMode = require("./text").Mode;
 var JsxHighlightRules = require("./jsx_highlight_rules").JsxHighlightRules;
 var MatchingBraceOutdent = require("./matching_brace_outdent").MatchingBraceOutdent;
-var DocCommentBehaviour = require("../mode/behaviour/doc_comment").DocCommentBehaviour;
+var CstyleBehaviour = require("../mode/behaviour/cstyle").CstyleBehaviour;
 var CStyleFoldMode = require("./folding/cstyle").FoldMode;
 
 function Mode() {
     this.HighlightRules = JsxHighlightRules;
     this.$outdent = new MatchingBraceOutdent();
-    this.$behaviour = new DocCommentBehaviour();
+    this.$behaviour = new CstyleBehaviour({closeDocComment: true});
     this.foldingRules = new CStyleFoldMode();
 }
 oop.inherits(Mode, TextMode);
