+++ src/mode/rust.js
@@ -8,12 +8,12 @@ var oop = require("../lib/oop");
 var TextMode = require("./text").Mode;
 var RustHighlightRules = require("./rust_highlight_rules").RustHighlightRules;
 var FoldMode = require("./folding/cstyle").FoldMode;
-var DocCommentBehaviour = require("../mode/behaviour/doc_comment").DocCommentBehaviour;
+var CstyleBehaviour = require("../mode/behaviour/cstyle").CstyleBehaviour;
 
 var Mode = function() {
     this.HighlightRules = RustHighlightRules;
     this.foldingRules = new FoldMode();
-    this.$behaviour = new DocCommentBehaviour();
+    this.$behaviour = new CstyleBehaviour({closeDocComment: true});
 };
 oop.inherits(Mode, TextMode);
