+++ src/mode/sac.js
@@ -4,12 +4,12 @@ var oop = require("../lib/oop");
 var TextMode = require("./text").Mode;
 var SaCHighlightRules = require("./sac_highlight_rules").sacHighlightRules;
 var FoldMode = require("./folding/cstyle").FoldMode;
-var DocCommentBehaviour = require("../mode/behaviour/doc_comment").DocCommentBehaviour;
+var CstyleBehaviour = require("../mode/behaviour/cstyle").CstyleBehaviour;
 
 var Mode = function() {
   this.HighlightRules = SaCHighlightRules;
   this.foldingRules = new FoldMode();
-  this.$behaviour = new DocCommentBehaviour();
+  this.$behaviour = new CstyleBehaviour({closeDocComment: true});
 };
 oop.inherits(Mode, TextMode);
