+++ src/mode/scrypt.js
@@ -8,12 +8,12 @@ var oop = require("../lib/oop");
 var TextMode = require("./text").Mode;
 var scryptHighlightRules = require("./scrypt_highlight_rules").scryptHighlightRules;
 var FoldMode = require("./folding/cstyle").FoldMode;
-var DocCommentBehaviour = require("../mode/behaviour/doc_comment").DocCommentBehaviour;
+var CstyleBehaviour = require("../mode/behaviour/cstyle").CstyleBehaviour;
 
 var Mode = function () {
     this.HighlightRules = scryptHighlightRules;
     this.foldingRules = new FoldMode();
-    this.$behaviour = new DocCommentBehaviour();
+    this.$behaviour = new CstyleBehaviour({closeDocComment: true});
 };
 oop.inherits(Mode, TextMode);
